Split full_node_api.py: extract wallet protocol handlers into mixin
The FullNodeAPI class (2070 lines) was too large to easily understand
and modify. This splits it into two files by extracting all wallet
protocol handlers into a separate mixin class:

- chia/full_node/_full_node_api_wallet.py (814 lines):
  Contains _FullNodeApiWalletMixin with all wallet protocol methods
  (request_block_header, request_additions, request_removals,
  send_transaction, request_puzzle_solution, etc.) plus wallet-related
  helper methods (mempool_updates, subscriptions, is_trusted).

- chia/full_node/full_node_api.py (1294 lines):
  Retains FullNodeAPI class definition, peer protocol, full node
  protocol, farmer protocol, timelord protocol, and compact VDF
  handlers. Now inherits from _FullNodeApiWalletMixin.

The split uses a shared ApiMetadata instance (full_node_metadata) so
that @metadata.request() decorators in both the mixin and main class
register handlers into the same registry. All 51 API methods remain
accessible on FullNodeAPI, and all existing imports continue to work.

No functional changes.
